게시글 등록
게시글 등록 기능

+++ src/main/java/com/lec/packages/domain/primaryKeyClasses/ClubBoardKeyClass.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Getter
 public class ClubBoardKeyClass implements Serializable{
     private String clubCode;
-    private Long boardNo;
+    private int boardNo;
 }
